pyproject.toml
[tool.poetry]
name = "c-plus-product-register-flask"
version = "0.1.0"
description = ""
authors = ["Jhelison Uchoa <68653689+jhelison@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
fdb = "^2.0.1"
Flask-RESTful = "^0.3.9"
python-decouple = "^3.4"
Flask-SQLAlchemy = "^2.5.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
